Adicionando mensagens de alerta.

+++ src/components/Test.tsx
@@ -6,6 +6,7 @@ import { Table, Modal } from 'react-bootstrap'
 import Moment from 'moment'
 import { BsPencil } from 'react-icons/bs'
 import { VscChromeClose } from 'react-icons/vsc'
+import swal from 'sweetalert'
 
 interface JSONPlaceholder {
     id: number;
@@ -25,7 +26,7 @@ export default function Test() {
     const [modalDefault, setModalDefaultShow] = useState(false);
     const modalDefaultClose = () => setModalDefaultShow(false);
     const modalDefaultShow = () => setModalDefaultShow(true);
-    
+
     const [placeholders, setPlaceholder] = useState<JSONPlaceholder[]>([]);
     const [flexsAll, setFlexAll] = useState<Flex[]>([]);
 
@@ -69,7 +70,12 @@ export default function Test() {
             headers: {
                 'Content-type': 'application/json; charset=UTF-8',
             },
-        })
+        }).then(() => swal({
+            title: "Cadastrado com Sucesso!!!",
+            icon: "success",
+            buttons: [false],
+            timer: 2000,
+        }))
     };
 
     // ALTERAR - GET/:id
@@ -100,14 +106,39 @@ export default function Test() {
 
         fetch(`https://provadev.xlab.digital/api/v1/divida/${idPut}?uuid=8d7297ad-3caa-4bab-9e16-99653958fac5`, requestOptions)
             .then(response => response.json())
+            .then(() => swal({
+                title: "Alterado com Sucesso!!!",
+                icon: "success",
+                buttons: [false],
+                timer: 2000,
+            }))
+
+        modalDefaultClose()
     };
 
     // DELETAR - DELETE/:id
     async function handleDelete(id: string) {
+        swal({
+            title: "Tem certeza?",
+            text: "Uma vez excluído, você não será capaz de recuperar este arquivo!",
+            icon: "warning",
+            buttons: ["Cancelar", true],
+            dangerMode: true,
+        }).then((willDelete) => {
+            if (willDelete) {
+                fetch(`https://provadev.xlab.digital/api/v1/divida/${id}?uuid=8d7297ad-3caa-4bab-9e16-99653958fac5`, {
+                    method: 'DELETE',
+                });
+                swal({
+                    title: "Excluído com Sucesso!!!",
+                    icon: "success",
+                    buttons: [false],
+                    timer: 2000,
+                });
+            } else {
 
-        fetch(`https://provadev.xlab.digital/api/v1/divida/${id}?uuid=8d7297ad-3caa-4bab-9e16-99653958fac5`, {
-            method: 'DELETE',
-        });
+            }
+        })
     }
 
     return (
